refactor: revise Recruit page
